+++ firebase-crashlytics/src/androidTest/java/com/google/firebase/crashlytics/internal/common/CrashlyticsControllerTest.java
@@ -34,13 +34,11 @@ import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.TaskCompletionSource;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApp;
-import com.google.firebase.concurrent.TestOnlyExecutors;
 import com.google.firebase.crashlytics.internal.CrashlyticsNativeComponent;
 import com.google.firebase.crashlytics.internal.CrashlyticsTestCase;
 import com.google.firebase.crashlytics.internal.DevelopmentPlatformProvider;
 import com.google.firebase.crashlytics.internal.NativeSessionFileProvider;
 import com.google.firebase.crashlytics.internal.analytics.AnalyticsEventLogger;
-import com.google.firebase.crashlytics.internal.concurrency.CrashlyticsWorkers;
 import com.google.firebase.crashlytics.internal.metadata.LogFileManager;
 import com.google.firebase.crashlytics.internal.metadata.UserMetadata;
 import com.google.firebase.crashlytics.internal.model.CrashlyticsReport;
@@ -57,14 +55,12 @@ import java.util.List;
 import java.util.TreeSet;
 import java.util.concurrent.Executor;
 import java.util.concurrent.TimeUnit;
+import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 
 public class CrashlyticsControllerTest extends CrashlyticsTestCase {
   private static final String GOOGLE_APP_ID = "google:app:id";
   private static final String SESSION_ID = "session_id";
-
-  private final CrashlyticsWorkers crashlyticsWorkers =
-      new CrashlyticsWorkers(TestOnlyExecutors.background(), TestOnlyExecutors.blocking());
 
   private Context testContext;
   private IdManager idManager;
@@ -103,12 +99,6 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
     when(testSettingsProvider.getSettingsAsync()).thenReturn(Tasks.forResult(testSettings));
   }
 
-  @Override
-  protected void tearDown() throws Exception {
-    super.tearDown();
-    crashlyticsWorkers.common.await();
-  }
-
   /** A convenience class for building CrashlyticsController instances for testing. */
   private class ControllerBuilder {
     private DataCollectionArbiter dataCollectionArbiter;
@@ -116,6 +106,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
     private AnalyticsEventLogger analyticsEventLogger;
     private SessionReportingCoordinator sessionReportingCoordinator;
 
+    private CrashlyticsBackgroundWorker backgroundWorker;
     private LogFileManager logFileManager = null;
 
     private UserMetadata userMetadata = null;
@@ -127,6 +118,8 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
       analyticsEventLogger = mock(AnalyticsEventLogger.class);
 
       sessionReportingCoordinator = mockSessionReportingCoordinator;
+
+      backgroundWorker = new CrashlyticsBackgroundWorker(new SameThreadExecutorService());
     }
 
     ControllerBuilder setDataCollectionArbiter(DataCollectionArbiter arbiter) {
@@ -175,6 +168,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
       final CrashlyticsController controller =
           new CrashlyticsController(
               testContext.getApplicationContext(),
+              backgroundWorker,
               idManager,
               dataCollectionArbiter,
               testFileStore,
@@ -185,8 +179,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
               sessionReportingCoordinator,
               nativeComponent,
               analyticsEventLogger,
-              mock(CrashlyticsAppQualitySessionsSubscriber.class),
-              crashlyticsWorkers);
+              mock(CrashlyticsAppQualitySessionsSubscriber.class));
       return controller;
     }
   }
@@ -213,10 +206,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
         .thenReturn(new TreeSet<>(Collections.singleton(sessionId)));
 
     controller.writeNonFatalException(thread, nonFatal);
-    crashlyticsWorkers.common.submit(() -> controller.doCloseSessions(testSettingsProvider));
-
-    crashlyticsWorkers.common.await();
-    crashlyticsWorkers.diskWrite.await();
+    controller.doCloseSessions(testSettingsProvider);
 
     verify(mockSessionReportingCoordinator)
         .persistNonFatalEvent(eq(nonFatal), eq(thread), eq(sessionId), anyLong());
@@ -238,8 +228,9 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
         .persistFatalEvent(eq(fatal), eq(thread), eq(sessionId), anyLong());
   }
 
+  @Test
   @SdkSuppress(minSdkVersion = 30) // ApplicationExitInfo
-  public void testOnDemandFatal_callLogFatalException() throws Exception {
+  public void testOnDemandFatal_callLogFatalException() {
     Thread thread = Thread.currentThread();
     Exception fatal = new RuntimeException("Fatal");
     Thread.UncaughtExceptionHandler exceptionHandler = mock(Thread.UncaughtExceptionHandler.class);
@@ -254,8 +245,6 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
             .build();
     controller.enableExceptionHandling(SESSION_ID, exceptionHandler, testSettingsProvider);
     controller.logFatalException(thread, fatal);
-
-    crashlyticsWorkers.common.await();
 
     verify(mockUserMetadata).setNewSession(not(eq(SESSION_ID)));
   }
@@ -334,9 +323,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
     final CrashlyticsController controller =
         builder().setNativeComponent(mockNativeComponent).setLogFileManager(logFileManager).build();
 
-    crashlyticsWorkers.common.submit(() -> controller.finalizeSessions(testSettingsProvider));
-    crashlyticsWorkers.common.await();
-
+    controller.finalizeSessions(testSettingsProvider);
     verify(mockSessionReportingCoordinator)
         .finalizeSessionWithNativeEvent(eq(previousSessionId), any(), any());
     verify(mockSessionReportingCoordinator, never())
@@ -344,10 +331,9 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
   }
 
   @SdkSuppress(minSdkVersion = 30) // ApplicationExitInfo
-  public void testMissingNativeComponentCausesNoReports() throws Exception {
+  public void testMissingNativeComponentCausesNoReports() {
     final CrashlyticsController controller = createController();
-    crashlyticsWorkers.common.submit(() -> controller.finalizeSessions(testSettingsProvider));
-    crashlyticsWorkers.common.await();
+    controller.finalizeSessions(testSettingsProvider);
 
     List<String> sessions = testFileStore.getAllOpenSessionIds();
     for (String sessionId : sessions) {
@@ -377,15 +363,13 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
    */
   // FIXME: Validate this test works as intended
   @SdkSuppress(minSdkVersion = 30) // ApplicationExitInfo
-  public void testLogStringAfterCrashOk() throws Exception {
+  public void testLogStringAfterCrashOk() {
     final CrashlyticsController controller = builder().build();
     controller.handleUncaughtException(
         testSettingsProvider, Thread.currentThread(), new RuntimeException());
 
     // This should not throw.
-    crashlyticsWorkers.diskWrite.submit(
-        () -> controller.writeToLog(System.currentTimeMillis(), "Hi"));
-    crashlyticsWorkers.diskWrite.await();
+    controller.writeToLog(System.currentTimeMillis(), "Hi");
   }
 
   /**
@@ -400,8 +384,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
         testSettingsProvider, Thread.currentThread(), new RuntimeException());
 
     // This should not throw.
-    crashlyticsWorkers.common.submit(() -> controller.finalizeSessions(testSettingsProvider));
-    crashlyticsWorkers.common.await();
+    controller.finalizeSessions(testSettingsProvider);
   }
 
   @SdkSuppress(minSdkVersion = 30) // ApplicationExitInfo
@@ -444,7 +427,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
 
     final DataCollectionArbiter arbiter = mock(DataCollectionArbiter.class);
     when(arbiter.isAutomaticDataCollectionEnabled()).thenReturn(false);
-    when(arbiter.waitForDataCollectionPermission())
+    when(arbiter.waitForDataCollectionPermission(any(Executor.class)))
         .thenReturn(new TaskCompletionSource<Void>().getTask());
     when(arbiter.waitForAutomaticDataCollectionEnabled())
         .thenReturn(new TaskCompletionSource<Void>().getTask());
@@ -489,8 +472,6 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
 
     await(controller.deleteUnsentReports());
     await(task);
-
-    crashlyticsWorkers.diskWrite.await();
 
     verify(mockSessionReportingCoordinator).removeAllReports();
     verifyNoMoreInteractions(mockSessionReportingCoordinator);
@@ -554,7 +535,7 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
   }
 
   @SdkSuppress(minSdkVersion = 30) // ApplicationExitInfo
-  public void testFatalEvent_sendsAppExceptionEvent() throws Exception {
+  public void testFatalEvent_sendsAppExceptionEvent() {
     final String sessionId = "sessionId";
     final LogFileManager logFileManager = new LogFileManager(testFileStore);
     final AnalyticsEventLogger mockFirebaseAnalyticsLogger = mock(AnalyticsEventLogger.class);
@@ -567,14 +548,10 @@ public class CrashlyticsControllerTest extends CrashlyticsTestCase {
     when(mockSessionReportingCoordinator.listSortedOpenSessionIds())
         .thenReturn(new TreeSet<>(Collections.singleton(sessionId)));
 
-    crashlyticsWorkers.common.submit(
-        () -> {
-          controller.openSession(SESSION_ID);
-          controller.handleUncaughtException(
-              testSettingsProvider, Thread.currentThread(), new RuntimeException("Fatal"));
-          controller.finalizeSessions(testSettingsProvider);
-        });
-    crashlyticsWorkers.common.await();
+    controller.openSession(SESSION_ID);
+    controller.handleUncaughtException(
+        testSettingsProvider, Thread.currentThread(), new RuntimeException("Fatal"));
+    controller.finalizeSessions(testSettingsProvider);
 
     assertFirebaseAnalyticsCrashEvent(mockFirebaseAnalyticsLogger);
   }