+++ firebase-crashlytics/src/androidTest/java/com/google/firebase/crashlytics/internal/common/SessionReportingCoordinatorTest.java
@@ -33,8 +33,6 @@ import static org.mockito.Mockito.when;
 
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
-import com.google.firebase.concurrent.TestOnlyExecutors;
-import com.google.firebase.crashlytics.internal.concurrency.CrashlyticsWorkers;
 import com.google.firebase.crashlytics.internal.metadata.LogFileManager;
 import com.google.firebase.crashlytics.internal.metadata.UserMetadata;
 import com.google.firebase.crashlytics.internal.model.CrashlyticsReport;
@@ -71,9 +69,6 @@ public class SessionReportingCoordinatorTest {
 
   private SessionReportingCoordinator reportingCoordinator;
 
-  private CrashlyticsWorkers crashlyticsWorkers =
-      new CrashlyticsWorkers(TestOnlyExecutors.background(), TestOnlyExecutors.blocking());
-
   @Before
   public void setUp() {
     MockitoAnnotations.initMocks(this);
@@ -85,8 +80,7 @@ public class SessionReportingCoordinatorTest {
             reportSender,
             logFileManager,
             reportMetadata,
-            idManager,
-            crashlyticsWorkers);
+            idManager);
   }
 
   @Test
@@ -122,8 +116,7 @@ public class SessionReportingCoordinatorTest {
   }
 
   @Test
-  public void testNonFatalEvent_persistsNormalPriorityEventWithoutAllThreadsForSessionId()
-      throws Exception {
+  public void testNonFatalEvent_persistsNormalPriorityEventWithoutAllThreadsForSessionId() {
     final String eventType = "error";
     final String sessionId = "testSessionId";
     final long timestamp = System.currentTimeMillis();
@@ -132,8 +125,6 @@ public class SessionReportingCoordinatorTest {
 
     reportingCoordinator.onBeginSession(sessionId, timestamp);
     reportingCoordinator.persistNonFatalEvent(mockException, mockThread, sessionId, timestamp);
-
-    crashlyticsWorkers.diskWrite.await();
 
     final boolean expectedAllThreads = false;
     final boolean expectedHighPriority = false;
@@ -145,7 +136,7 @@ public class SessionReportingCoordinatorTest {
   }
 
   @Test
-  public void testNonFatalEvent_addsLogsToEvent() throws Exception {
+  public void testNonFatalEvent_addsLogsToEvent() {
     long timestamp = System.currentTimeMillis();
 
     mockEventInteractions();
@@ -158,8 +149,6 @@ public class SessionReportingCoordinatorTest {
     reportingCoordinator.onBeginSession(sessionId, timestamp);
     reportingCoordinator.persistNonFatalEvent(mockException, mockThread, sessionId, timestamp);
 
-    crashlyticsWorkers.diskWrite.await();
-
     verify(mockEventBuilder)
         .setLog(CrashlyticsReport.Session.Event.Log.builder().setContent(testLog).build());
     verify(mockEventBuilder).build();
@@ -167,7 +156,7 @@ public class SessionReportingCoordinatorTest {
   }
 
   @Test
-  public void testNonFatalEvent_addsNoLogsToEventWhenNoneAvailable() throws Exception {
+  public void testNonFatalEvent_addsNoLogsToEventWhenNoneAvailable() {
     long timestamp = System.currentTimeMillis();
 
     mockEventInteractions();
@@ -178,8 +167,6 @@ public class SessionReportingCoordinatorTest {
 
     reportingCoordinator.onBeginSession(sessionId, timestamp);
     reportingCoordinator.persistNonFatalEvent(mockException, mockThread, sessionId, timestamp);
-
-    crashlyticsWorkers.diskWrite.await();
 
     verify(mockEventBuilder, never()).setLog(any(CrashlyticsReport.Session.Event.Log.class));
     verify(mockEventBuilder).build();
@@ -225,7 +212,7 @@ public class SessionReportingCoordinatorTest {
   }
 
   @Test
-  public void testNonFatalEvent_addsSortedKeysToEvent() throws Exception {
+  public void testNonFatalEvent_addsSortedKeysToEvent() {
     final long timestamp = System.currentTimeMillis();
 
     mockEventInteractions();
@@ -256,8 +243,6 @@ public class SessionReportingCoordinatorTest {
     reportingCoordinator.onBeginSession(sessionId, timestamp);
     reportingCoordinator.persistNonFatalEvent(mockException, mockThread, sessionId, timestamp);
 
-    crashlyticsWorkers.diskWrite.await();
-
     verify(mockEventAppBuilder).setCustomAttributes(expectedCustomAttributes);
     verify(mockEventAppBuilder).setInternalKeys(expectedCustomAttributes);
     verify(mockEventAppBuilder).build();
@@ -267,7 +252,7 @@ public class SessionReportingCoordinatorTest {
   }
 
   @Test
-  public void testNonFatalEvent_addsNoKeysToEventWhenNoneAvailable() throws Exception {
+  public void testNonFatalEvent_addsNoKeysToEventWhenNoneAvailable() {
     final long timestamp = System.currentTimeMillis();
 
     mockEventInteractions();
@@ -281,8 +266,6 @@ public class SessionReportingCoordinatorTest {
     reportingCoordinator.onBeginSession(sessionId, timestamp);
     reportingCoordinator.persistNonFatalEvent(mockException, mockThread, sessionId, timestamp);
 
-    crashlyticsWorkers.diskWrite.await();
-
     verify(mockEventAppBuilder, never()).setCustomAttributes(anyList());
     verify(mockEventAppBuilder, never()).build();
     verify(mockEventBuilder, never()).setApp(mockEventApp);
@@ -291,7 +274,7 @@ public class SessionReportingCoordinatorTest {
   }
 
   @Test
-  public void testNonFatalEvent_addRolloutsEvent() throws Exception {
+  public void testNonFatalEvent_addRolloutsEvent() {
     long timestamp = System.currentTimeMillis();
     String sessionId = "testSessionId";
     mockEventInteractions();
@@ -303,8 +286,6 @@ public class SessionReportingCoordinatorTest {
 
     reportingCoordinator.onBeginSession(sessionId, timestamp);
     reportingCoordinator.persistNonFatalEvent(mockException, mockThread, sessionId, timestamp);
-
-    crashlyticsWorkers.diskWrite.await();
 
     verify(mockEventAppBuilder, never()).setCustomAttributes(anyList());
     verify(mockEventAppBuilder, never()).build();
@@ -434,6 +415,35 @@ public class SessionReportingCoordinatorTest {
     // first build for custom keys
     // second build for rollouts
     verify(mockEventBuilder, times(2)).build();
+  }
+
+  @Test
+  public void onLog_writesToLogFileManager() {
+    long timestamp = System.currentTimeMillis();
+    String log = "this is a log";
+
+    reportingCoordinator.onLog(timestamp, log);
+
+    verify(logFileManager).writeToLog(timestamp, log);
+  }
+
+  @Test
+  public void onCustomKey_writesToReportMetadata() {
+    final String key = "key";
+    final String value = "value";
+
+    reportingCoordinator.onCustomKey(key, value);
+
+    verify(reportMetadata).setCustomKey(key, value);
+  }
+
+  @Test
+  public void onUserId_writesUserToReportMetadata() {
+    final String userId = "testUser";
+
+    reportingCoordinator.onUserId(userId);
+
+    verify(reportMetadata).setUserId(userId);
   }
 
   @Test