+++ firebase-crashlytics/src/main/java/com/google/firebase/crashlytics/FirebaseCrashlytics.java
@@ -19,9 +19,12 @@ import android.content.pm.PackageManager;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
+import com.google.android.gms.tasks.Continuation;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.analytics.connector.AnalyticsConnector;
+import com.google.firebase.annotations.concurrent.Background;
+import com.google.firebase.annotations.concurrent.Blocking;
 import com.google.firebase.crashlytics.internal.CrashlyticsNativeComponent;
 import com.google.firebase.crashlytics.internal.CrashlyticsNativeComponentDeferredProxy;
 import com.google.firebase.crashlytics.internal.DevelopmentPlatformProvider;
@@ -33,8 +36,8 @@ import com.google.firebase.crashlytics.internal.common.CommonUtils;
 import com.google.firebase.crashlytics.internal.common.CrashlyticsAppQualitySessionsSubscriber;
 import com.google.firebase.crashlytics.internal.common.CrashlyticsCore;
 import com.google.firebase.crashlytics.internal.common.DataCollectionArbiter;
+import com.google.firebase.crashlytics.internal.common.ExecutorUtils;
 import com.google.firebase.crashlytics.internal.common.IdManager;
-import com.google.firebase.crashlytics.internal.concurrency.CrashlyticsWorkers;
 import com.google.firebase.crashlytics.internal.network.HttpRequestFactory;
 import com.google.firebase.crashlytics.internal.persistence.FileStore;
 import com.google.firebase.crashlytics.internal.settings.SettingsController;
@@ -43,6 +46,7 @@ import com.google.firebase.installations.FirebaseInstallationsApi;
 import com.google.firebase.remoteconfig.interop.FirebaseRemoteConfigInterop;
 import com.google.firebase.sessions.api.FirebaseSessionsDependencies;
 import java.util.List;
+import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 
 /**
@@ -66,8 +70,8 @@ public class FirebaseCrashlytics {
       @NonNull Deferred<CrashlyticsNativeComponent> nativeComponent,
       @NonNull Deferred<AnalyticsConnector> analyticsConnector,
       @NonNull Deferred<FirebaseRemoteConfigInterop> remoteConfigInteropDeferred,
-      ExecutorService backgroundExecutorService,
-      ExecutorService blockingExecutorService) {
+      @Background ExecutorService backgroundExecutorService,
+      @Blocking ExecutorService blockingExecutorService) {
 
     Context context = app.getApplicationContext();
     final String appIdentifier = context.getPackageName();
@@ -77,9 +81,6 @@ public class FirebaseCrashlytics {
                 + CrashlyticsCore.getVersion()
                 + " for "
                 + appIdentifier);
-
-    CrashlyticsWorkers crashlyticsWorkers =
-        new CrashlyticsWorkers(backgroundExecutorService, blockingExecutorService);
 
     FileStore fileStore = new FileStore(context);
     final DataCollectionArbiter arbiter = new DataCollectionArbiter(app);
@@ -91,6 +92,9 @@ public class FirebaseCrashlytics {
     // Integration with Firebase Analytics
     final AnalyticsDeferredProxy analyticsDeferredProxy =
         new AnalyticsDeferredProxy(analyticsConnector);
+
+    final ExecutorService crashHandlerExecutor =
+        ExecutorUtils.buildSingleThreadExecutorService("Crashlytics Exception Handler");
 
     CrashlyticsAppQualitySessionsSubscriber sessionsSubscriber =
         new CrashlyticsAppQualitySessionsSubscriber(arbiter, fileStore);
@@ -108,9 +112,9 @@ public class FirebaseCrashlytics {
             analyticsDeferredProxy.getDeferredBreadcrumbSource(),
             analyticsDeferredProxy.getAnalyticsEventLogger(),
             fileStore,
+            crashHandlerExecutor,
             sessionsSubscriber,
-            remoteConfigDeferredProxy,
-            crashlyticsWorkers);
+            remoteConfigDeferredProxy);
 
     final String googleAppId = app.getOptions().getApplicationId();
     final String mappingFileId = CommonUtils.getMappingFileId(context);
@@ -146,7 +150,10 @@ public class FirebaseCrashlytics {
 
     Logger.getLogger().v("Installer package name is: " + appData.installerPackageName);
 
-    SettingsController settingsController =
+    final Executor threadPoolExecutor =
+        ExecutorUtils.buildSequentialExecutor(backgroundExecutorService);
+
+    final SettingsController settingsController =
         SettingsController.create(
             context,
             googleAppId,
@@ -159,8 +166,18 @@ public class FirebaseCrashlytics {
 
     // Kick off actually fetching the settings.
     settingsController
-        .loadSettingsData(crashlyticsWorkers)
-        .addOnFailureListener(ex -> Logger.getLogger().e("Error fetching settings.", ex));
+        .loadSettingsData(threadPoolExecutor)
+        .continueWith(
+            threadPoolExecutor,
+            new Continuation<Void, Object>() {
+              @Override
+              public Object then(@NonNull Task<Void> task) throws Exception {
+                if (!task.isSuccessful()) {
+                  Logger.getLogger().e("Error fetching settings.", task.getException());
+                }
+                return null;
+              }
+            });
 
     final boolean finishCoreInBackground = core.onPreExecute(appData, settingsController);
 
@@ -439,11 +456,11 @@ public class FirebaseCrashlytics {
    * Indicates whether or not automatic data collection is enabled
    *
    * @return In order of priority:
-   *     <p>If {@link #setCrashlyticsCollectionEnabled(boolean)} is called with a value, use it
-   *     <p>If the <b>firebase_crashlytics_collection_enabled</b> key is in your app’s
-   *     AndroidManifest.xml, use it
-   *     <p>Otherwise, use the default {@link FirebaseApp#isDataCollectionDefaultEnabled()} in
-   *     FirebaseApp
+   * <p>If {@link #setCrashlyticsCollectionEnabled(boolean)} is called with a value, use it</p>
+   *
+   * <p>If the <b>firebase_crashlytics_collection_enabled</b> key is in your app’s AndroidManifest.xml, use it</p>
+   *
+   * <p>Otherwise, use the default {@link FirebaseApp#isDataCollectionDefaultEnabled()} in FirebaseApp</p>
    */
   public boolean isCrashlyticsCollectionEnabled() {
     return core.isCrashlyticsCollectionEnabled();